+++ example/lib/live_decode.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+import 'package:qr_code_dart_scan/qr_code_dart_scan.dart';
+
+///
+/// Created by
+///
+/// ─▄▀─▄▀
+/// ──▀──▀
+/// █▀▀▀▀▀█▄
+/// █░░░░░█─█
+/// ▀▄▄▄▄▄▀▀
+///
+/// Rafaelbarbosatec
+/// on 28/06/22
+class LiveDecodePage extends StatefulWidget {
+  static const routeName = '/live';
+  static get route => {routeName: (BuildContext context) => LiveDecodePage()};
+  static open(BuildContext context) {
+    Navigator.of(context).pushNamed(routeName);
+  }
+
+  LiveDecodePage({Key? key}) : super(key: key);
+
+  @override
+  _LiveDecodePageState createState() => _LiveDecodePageState();
+}
+
+class _LiveDecodePageState extends State<LiveDecodePage> {
+  Result? currentResult;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: QRCodeDartScanView(
+        scanInvertedQRCode: true,
+        onCapture: (Result result) {
+          setState(() {
+            currentResult = result;
+          });
+        },
+        child: Align(
+          alignment: Alignment.bottomCenter,
+          child: Container(
+            margin: EdgeInsets.all(20),
+            padding: EdgeInsets.all(20),
+            decoration: BoxDecoration(
+              color: Colors.white,
+              borderRadius: BorderRadius.circular(20),
+            ),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text('Text: ${currentResult?.text ?? 'Not found'}'),
+                Text('Format: ${currentResult?.barcodeFormat ?? 'Not found'}'),
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
