+++ lib/qr_code_dart_scan.dart
@@ -3,4 +3,4 @@ library qr_code_dart_scan;
 export 'package:zxing_lib/zxing.dart';
 
 export 'src/qr_code_dart_scan_view.dart';
-export 'src/util.dart';
+export 'src/util/qr_code_dart_scan_resolution_preset.dart';
