+++ lib/src/decoder/qr_code_dart_scan_multi_reader.dart
