+++ lib/src/util/qr_code_dart_scan_resolution_preset.dart
